% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4h_classify_rows}
\alias{s4h_classify_rows}
\title{Classify Rows of a Data Frame with the BERT Model}
\usage{
s4h_classify_rows(
  data,
  col1,
  col2,
  col3,
  new_column_name = "category",
  MODEL_PATH = "./bert_finetuned_classifier"
)
}
\arguments{
\item{data}{R data.frame with text columns.}

\item{col1, col2, col3}{Names of the text columns.}

\item{new_column_name}{Name of the new column with the category.}

\item{MODEL_PATH}{Path to the model (defaults to \code{"./bert_finetuned_classifier"}).}
}
\value{
R data.frame with a new prediction column.
}
\description{
Wrapper for \code{s4h_classify_rows()} in Python.
}
