% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_compare_with_dict}
\alias{s4h_compare_with_dict}
\title{Compare Dask DataFrame Columns Against Dictionary}
\usage{
s4h_compare_with_dict(harmonizer, ddfs, as_data_frame = TRUE)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object with \code{dict_df} defined.}

\item{ddfs}{List of Dask DataFrames.}

\item{as_data_frame}{\code{TRUE} to return an R data.frame.}
}
\value{
pandas DataFrame (Python) or R data.frame.
}
\description{
Wrapper for \code{Harmonizer.s4h_compare_with_dict()}.
}
