% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{s4h_delete_download_folder}
\alias{s4h_delete_download_folder}
\title{Delete the Download Folder from an \code{Extractor} Object}
\usage{
s4h_delete_download_folder(extractor, folder_path = NULL)
}
\arguments{
\item{extractor}{Python \code{Extractor} object.}

\item{folder_path}{Specific path to delete (optional).}
}
\value{
\code{TRUE}/\code{FALSE} based on the result from Python.
}
\description{
Wrapper for \code{Extractor.s4h_delete_download_folder()}.
}
