% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_drop_nan_columns}
\alias{s4h_drop_nan_columns}
\title{Drop Columns with Many NaN Values using \code{Harmonizer}}
\usage{
s4h_drop_nan_columns(
  harmonizer,
  ddf_or_ddfs,
  return_as = c("dask", "pandas", "data.frame")
)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object.}

\item{ddf_or_ddfs}{A Dask DataFrame or list of Dask DataFrames.}

\item{return_as}{\code{"dask"}, \code{"pandas"}, or \code{"data.frame"}.}
}
\value{
Object(s) in the specified format.
}
\description{
Wrapper for \code{Harmonizer.drop_nan_columns()}.
}
