% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{s4h_extract}
\alias{s4h_extract}
\title{Execute \code{s4h_extract()} from an \code{Extractor} Object}
\usage{
s4h_extract(extractor, return_as = c("dask", "pandas", "data.frame"))
}
\arguments{
\item{extractor}{Python \code{Extractor} object (returned by \code{s4h_extractor()}).}

\item{return_as}{\code{"dask"}, \code{"pandas"}, or \code{"data.frame"}.}
}
\value{
\itemize{
\item \code{"dask"}: list of Dask objects (Python).
\item \code{"pandas"}: list of pandas DataFrames (Python).
\item \code{"data.frame"}: list of R data.frames.
}
}
\description{
Wrapper for \code{Extractor.s4h_extract()} in Python.
Depending on \code{return_as}, converts the list of Dask DataFrames to
pandas or R data.frames.
}
