% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_get_available_columns}
\alias{s4h_get_available_columns}
\title{Get Available Columns (Static)}
\usage{
s4h_get_available_columns(df_or_dfs)
}
\arguments{
\item{df_or_dfs}{Dask DataFrame, pandas DataFrame, or list of DataFrames.}
}
\value{
Vector of column names.
}
\description{
Wrapper for \code{Harmonizer.s4h_get_available_columns()}.
}
