% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4h_get_classifier}
\alias{s4h_get_classifier}
\title{Get (and Cache) the BERT Classifier}
\usage{
s4h_get_classifier(MODEL_PATH = "./bert_finetuned_classifier")
}
\arguments{
\item{MODEL_PATH}{Path to the model (or HuggingFace model identifier).}
}
\value{
\code{transformers.Pipeline} object (Python).
}
\description{
Wrapper for \code{s4h_get_classifier()} in Python.
}
