% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{s4h_get_default_data_dir}
\alias{s4h_get_default_data_dir}
\title{Get socio4health Default Data Directory}
\usage{
s4h_get_default_data_dir()
}
\value{
Character string with the path to the directory where socio4health stores downloads.
}
\description{
Wrapper for \code{s4h_get_default_data_dir()} Python function.
}
\examples{
\dontrun{
  data_dir <- s4h_get_default_data_dir()
  print(data_dir)
}

}
