% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_harmonize_dataframes}
\alias{s4h_harmonize_dataframes}
\title{Harmonize DataFrames by Country with \code{Harmonizer}}
\usage{
s4h_harmonize_dataframes(
  harmonizer,
  country_dfs,
  return_as = c("dask", "pandas", "data.frame")
)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object.}

\item{country_dfs}{Python dictionary or named list of lists of Dask DataFrames.}

\item{return_as}{\code{"dask"}, \code{"pandas"}, or \code{"data.frame"}.}
}
\value{
Equivalent structure (dictionary/named list) with harmonized DataFrames.
}
\description{
Wrapper for \code{Harmonizer.s4h_harmonize_dataframes()}.
}
