% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_harmonizer}
\alias{s4h_harmonizer}
\title{Create a socio4health \code{Harmonizer} Object}
\usage{
s4h_harmonizer(
  min_common_columns = 1,
  nan_threshold = 1,
  sample_frac = NULL,
  column_mapping = NULL,
  value_mappings = NULL,
  theme_info = NULL,
  default_country = NULL,
  strict_mapping = FALSE,
  dict_df = NULL,
  categories = NULL,
  key_col = NULL,
  key_val = NULL,
  extra_cols = NULL,
  join_key = NULL,
  aux_key = NULL
)
}
\arguments{
\item{min_common_columns}{Minimum number of common columns for vertical merge.}

\item{nan_threshold}{NaN threshold for removing columns between 0 and 1.}

\item{sample_frac}{Sampling fraction for NaN.}

\item{column_mapping}{Enum, dict, JSON, or path to JSON.}

\item{value_mappings}{Enum, dict, JSON, or path to JSON.}

\item{theme_info}{Dict/JSON/path with theme/category information.}

\item{default_country}{Default country.}

\item{strict_mapping}{\code{TRUE} to require complete mappings.}

\item{dict_df}{R data frame or pandas DataFrame with the dictionary.}

\item{categories}{Vector of categories.}

\item{key_col}{Key column for row selection.}

\item{key_val}{Key values.}

\item{extra_cols}{Extra columns to preserve.}

\item{join_key}{Key column for joins.}

\item{aux_key}{Auxiliary key for joins.}
}
\value{
Python \code{Harmonizer} object.
}
\description{
Direct wrapper for the \code{Harmonizer.__init__} constructor in Python.
All parameters are passed directly to the constructor.
}
