% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_harmonizer_call}
\alias{s4h_harmonizer_call}
\title{Call an Arbitrary Method on a \code{Harmonizer} Object}
\usage{
s4h_harmonizer_call(harmonizer, method, ...)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object.}

\item{method}{Name of the method in Python.}

\item{...}{Additional arguments.}
}
\value{
The return value of the Python method.
}
\description{
Call an Arbitrary Method on a \code{Harmonizer} Object
}
