% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_join_data}
\alias{s4h_join_data}
\title{Join Multiple Dask DataFrames with \code{Harmonizer.s4h_join_data()}}
\usage{
s4h_join_data(harmonizer, ddfs, as_data_frame = TRUE)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object (with \code{join_key} and, optionally, \code{aux_key}).}

\item{ddfs}{List of Dask DataFrames.}

\item{as_data_frame}{\code{TRUE} to return R data.frame.}
}
\value{
pandas DataFrame or R data.frame.
}
\description{
Join Multiple Dask DataFrames with \code{Harmonizer.s4h_join_data()}
}
