% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4h_parse_fwf_dict}
\alias{s4h_parse_fwf_dict}
\title{Parse FWF Dictionary with socio4health}
\usage{
s4h_parse_fwf_dict(dic)
}
\arguments{
\item{dic}{R data.frame that contains at least the columns:
\itemize{
\item variable_name
\item initial_position
\item size  or  final_position
}}
}
\value{
List with:
\itemize{
\item colnames: vector of column names
\item colspecs: list of pairs (start, end) in base 0 (end exclusive)
}
}
\description{
Wrapper for \code{s4h_parse_fwf_dict()} in
\code{socio4health.utils.extractor_utils}.
}
