% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4h_translate_column}
\alias{s4h_translate_column}
\title{Translate a Text Column with Google Translate}
\usage{
s4h_translate_column(data, column, language = "en")
}
\arguments{
\item{data}{R data.frame.}

\item{column}{Name of the column to translate.}

\item{language}{Target language code (e.g., \code{"en"}).}
}
\value{
R data.frame with a new translated column.
}
\description{
Translate a Text Column with Google Translate
}
