% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{s4h_check_env}
\alias{s4h_check_env}
\title{Check Python Environment Status}
\usage{
s4h_check_env()
}
\value{
Invisibly returns a list with environment information.
}
\description{
Displays information about the current Python environment and socio4health availability.
}
\examples{
\dontrun{
s4h_check_env()
}

}
