% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.R
\name{s4h_vertical_merge}
\alias{s4h_vertical_merge}
\title{Vertically Merge Dask DataFrames with \code{Harmonizer}}
\usage{
s4h_vertical_merge(
  harmonizer,
  ddfs,
  return_as = c("dask", "pandas", "data.frame")
)
}
\arguments{
\item{harmonizer}{Python \code{Harmonizer} object.}

\item{ddfs}{List of Dask DataFrames (typically output from \code{s4h_run_extract(..., "dask")}).}

\item{return_as}{\code{"dask"}, \code{"pandas"}, or \code{"data.frame"}.}
}
\value{
List of DataFrames in the specified format.
}
\description{
Wrapper for \code{Harmonizer.s4h_vertical_merge()}.
}
